relocates makeMove logic to its own file
